[feat][casey&pat] add tabs bar to dashboard
