--- src/components/Dashboard/styles.scss
@@ -1,76 +1,0 @@
-@import '../../utils/variables.scss';
-
-/* Dashboard Page */
-
-.dashboard-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.dashboard-button {
-  width: 250px;
-  height: 35px;
-  margin-top: 14px;
-
-  font-family: $ibm-plex-mono;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 0.85rem;
-  line-height: 16px;
-  /* identical to box height */
-  background-color: transparent;
-  border: 2px solid $white;
-  border-radius: 5px;
-  color: $white;
-
-  cursor: pointer;
-  /* Neutral/0 */
-}
-
-.dashboard-card {
-  /* dashboard Card */
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  width: 250px;
-  height: 90px;
-  box-shadow: $card-box-shadow;
-
-  background: $secondary-background;
-  border-radius: 16px;
-}
-
-.dashboard-margin {
-  margin: 2px;
-}
-
-.dashboard-text {
-  .ant-tabs-tab.ant-tabs-tab-active .ant-tabs-tab-btn {
-    color: $secondary-color;
-    text-shadow: 0 0 0.25px currentColor;
-  }
-  .ant-tabs-top > .ant-tabs-nav .ant-tabs-ink-bar, .ant-tabs-top > div > .ant-tabs-nav .ant-tabs-ink-bar {
-    bottom: 0;
-    color: $secondary-color;
-}
-  font-family: $ibm-plex-mono;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 0.85rem;
-  text-align: center;
-  color: $white;
-  margin-bottom: 10px;
-}
-
-.dashboard-subtitle {
-  align-self: auto;
-  font-family: $ibm-plex-mono;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 0.75rem;
-  text-align: center;
-  color: $white;
-  margin-bottom: 10px;
-}
