+++ src/pages/Dashboard/index.jsx
@@ -1,19 +1,34 @@
 import React from 'react';
+import { useHistory } from 'react-router-dom';
 
-import { Space } from 'antd';
+import { Button } from 'antd';
 
-import Dashboard from 'components/Dashboard';
+import TabsBar from 'components/TabsBar';
+
+import { useAuth } from 'hooks/Hooks';
+import routes from 'constants/routes';
 
 import './styles.scss';
 
 function DashboardPage() {
+  const { logout } = useAuth();
+  const history = useHistory();
+
+  async function handleLogout() {
+    try {
+      await logout();
+      history.push(routes.LOGIN);
+    } catch (error) {
+      // eslint-disable-next-line no-console
+      console.log(error.message);
+    }
+  }
   return (
     <div className="dashboard">
-      <div className="dashboard-page">
-        <Space align="center" direction="vertical" size="large">
-          <Dashboard />
-        </Space>
-      </div>
+      <TabsBar />
+      <Button onClick={handleLogout} type="primary" className="dashboard-button">
+        Logout
+      </Button>
     </div>
   );
 }
