+++ src/utils/variables.scss
@@ -2,6 +2,7 @@ $primary-background: #15322f;
 $secondary-background: #62a09d;
 $secondary-color: #8dc7c3;
 $white: #e0f0f5;
+$black: #313837;
 $card-box-shadow: 0px 0px 2px 2px rgba(208, 216, 243, 0.6);
 $card-background: linear-gradient(0deg, #262829, #262829), #fff6f6;
